@@ -1,6 +1,6 @@
 {
   "name": "localize-router",
-  "version": "0.7.0",
+  "version": "0.7.1",
   "description": "An implementation of routes localization for Angular 2",
   "scripts": {
     "test": "karma start",
@@ -39,48 +39,48 @@
     "@ngx-translate/core": "^6.0.0"
   },
   "devDependencies": {
-    "@angular/common": "2.4.8",
-    "@angular/compiler": "2.4.8",
-    "@angular/compiler-cli": "2.4.8",
-    "@angular/core": "2.4.8",
-    "@angular/http": "2.4.8",
-    "@angular/platform-browser": "2.4.8",
-    "@angular/platform-browser-dynamic": "2.4.8",
-    "@angular/platform-server": "2.4.8",
-    "@angular/router": "3.4.8",
+    "@angular/common": "~4.1.3",
+    "@angular/compiler": "~4.1.3",
+    "@angular/compiler-cli": "~4.1.3",
+    "@angular/core": "~4.1.3",
+    "@angular/http": "~4.1.3",
+    "@angular/platform-browser": "~4.1.3",
+    "@angular/platform-browser-dynamic": "~4.1.3",
+    "@angular/platform-server": "~4.1.3",
+    "@angular/router": "~4.1.3",
     "@types/es6-promise": "0.0.32",
-    "@types/hammerjs": "2.0.33",
-    "@types/jasmine": "2.5.41",
-    "@types/node": "7.0.4",
-    "awesome-typescript-loader": "3.0.0-beta.18",
-    "codelyzer": "2.0.0-beta.4",
-    "commitizen": "2.9.2",
+    "@types/hammerjs": "2.0.34",
+    "@types/jasmine": "2.5.47",
+    "@types/node": "7.0.22",
+    "awesome-typescript-loader": "3.1.3",
+    "codelyzer": "3.0.1",
+    "commitizen": "2.9.6",
     "concurrently": "3.1.0",
     "core-js": "2.4.1",
     "cz-conventional-changelog": "1.2.0",
-    "istanbul-instrumenter-loader": "0.2.0",
-    "jasmine-core": "2.5.2",
-    "karma": "1.3.0",
-    "karma-chrome-launcher": "2.0.0",
+    "istanbul-instrumenter-loader": "~2.0.0",
+    "jasmine-core": "2.6.2",
+    "karma": "1.7.0",
+    "karma-chrome-launcher": "2.1.1",
     "karma-coverage": "1.1.1",
-    "karma-jasmine": "1.0.2",
-    "karma-mocha-reporter": "^2.1.0",
-    "karma-remap-coverage": "^0.1.4",
+    "karma-jasmine": "1.1.0",
+    "karma-mocha-reporter": "2.2.3",
+    "karma-remap-coverage": "0.1.4",
     "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "2.0.2",
-    "loader-utils": "0.2.16",
+    "karma-webpack": "2.0.3",
+    "loader-utils": "1.1.0",
     "@ngx-translate/core": "6.0.0",
-    "reflect-metadata": "0.1.8",
-    "rxjs": "5.0.1",
-    "semantic-release": "4.3.5",
-    "source-map-loader": "0.1.5",
+    "reflect-metadata": "0.1.10",
+    "rxjs": "5.4.0",
+    "semantic-release": "6.3.6",
+    "source-map-loader": "0.2.1",
     "ts-helpers": "1.1.2",
-    "tslint": "4.0.2",
-    "tslint-loader": "3.3.0",
-    "typescript": "2.0.10",
-    "webpack": "2.2.1",
+    "tslint": "5.3.2",
+    "tslint-loader": "3.5.3",
+    "typescript": "2.3.3",
+    "webpack": "2.6.0",
     "webpack-node-externals": "^1.5.4",
-    "zone.js": "0.7.2"
+    "zone.js": "0.8.11"
   },
   "config": {
     "commitizen": {
