@@ -54,7 +54,7 @@ module.exports = function (config) {
     autoWatch: false,
 
     browsers: [
-      'Chrome'
+      'ChromeHeadless'
     ],
 
     customLaunchers: {
