@@ -8,7 +8,7 @@ import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { HomeModule } from './home/home.module';
-import { AboutModule } from './about/about.module';
+import {UsersModule} from "./users/users.module";
 
 export function HttpLoaderFactory(http: Http) {
   return new TranslateHttpLoader(http, '/assets/locales/', '.json');
@@ -31,7 +31,7 @@ export function HttpLoaderFactory(http: Http) {
     }),
     // feature modules
     HomeModule,
-    AboutModule,
+    UsersModule,
     AppRoutingModule
   ],
   providers: [],
